:art: General refactor of item detail
